chore(locks): simplify preliminary lock state
