+++ plt/plt-block-state/src/block_state.rs
@@ -3,8 +3,8 @@
 use crate::block_state::blob_store::{BackingStoreLoad, BackingStoreStore, DecodeError};
 use crate::block_state::external::{ExternalBlockStateOperations, ExternalBlockStateQuery};
 use crate::block_state::types::{
-    AccountWithCanonicalAddress, LockConfiguration, TokenAccountState, TokenConfiguration,
-    TokenIndex, TokenStateKey, TokenStateValue,
+    AccountWithCanonicalAddress, LockConfiguration, TokenAccountState, TokenAndAmount,
+    TokenConfiguration, TokenIndex, TokenStateKey, TokenStateValue,
 };
 use crate::block_state_interface::{
     BlockStateOperations, BlockStateQuery, OverflowError, RawTokenAmountDelta,
@@ -17,7 +17,7 @@ use concordium_base::constants::SHA256;
 use concordium_base::contracts_common::AccountAddress;
 use concordium_base::protocol_level_locks::LockId;
 use concordium_base::protocol_level_tokens::TokenId;
-use plt_scheduler_types::types::tokens::{RawTokenAmount, TokenAndAmount};
+use plt_scheduler_types::types::tokens::RawTokenAmount;
 use sha2::Digest;
 use std::collections::BTreeMap;
 
@@ -400,6 +400,10 @@ struct Lock {
     locked_balances: BTreeMap<AccountIndex, TokenAndAmount>,
     /// The configuration parameters for the lock.
     configuration: LockConfiguration,
-    /// The key-value state associated with the lock.
-    key_value_state: SimplisticTokenKeyValueState,
+    /// The state associated with the lock.
+    state: LockState,
 }
+
+/// The state of a single protocol-level lock.
+#[derive(Debug, Clone, Serialize)]
+struct LockState {}
